+++ DataModels/CheckTypeResult.cs
@@ -1,0 +1,9 @@
+
+namespace VidsNet
+{
+    public class CheckTypeResult {
+        public bool CorrectType {get;set;}
+        public bool WriteVirtualItem {get;set;}
+        public ItemType Type {get;set;}
+    }
+}