+++ DataModels/FrontEndItem.cs
@@ -1,0 +1,9 @@
+
+namespace VidsNet
+{
+    public class FrontEndItem {
+        public string Name {get;set;}
+        public int Parent {get;set;}
+    }
+
+}