+++ DataModels/Item.cs
@@ -6,7 +6,6 @@ namespace VidsNet
 
         public string Path { get; set; } 
         public int Id { get; set; }
-
         public ItemType Type {get;set;}
     }
 }