+++ Enums/ItemType.cs
@@ -2,6 +2,7 @@
 namespace VidsNet
 {
     public enum ItemType {
+        None = -1,
         Folder,
         Video,
         Subtitle