+++ Setup.cs
@@ -4,7 +4,7 @@ namespace VidsNet
     public static class Setup {
 
         public static bool UsersExist() {
-            using(var db = new DatabseContext()) {
+            using(var db = new DatabaseContext()) {
                 if(db.Users.Any()) {
                     return true;
                 }
@@ -13,7 +13,7 @@ namespace VidsNet
         }
 
         public static bool SettingsExist() {
-            using(var db = new DatabseContext()) {
+            using(var db = new DatabaseContext()) {
                 if(db.Settings.Any()) {
                     return true;
                 }